chained filters do not work.
At the moment we iterate over all transformations and apply the same operations
to each, ie either insert:1 or remove:1

However, with chains of transformations, the operations may differ between
transformations.  We need to transform those operations so that we apply the
correct operations.

For example, if we have two filters, we may need to transform insert:1 to noop
(if the item fails the first filter).

Generally we need to do something like the following:

  set operations to insert:1 or remove:1 (for addedItem, removedItem)
  loop over all but the last transformation, transforming the operations
  apply the transformed operations to the last transformation