Release version 0.3.0.
Version 0.3.0 is the result of a lot of work by both @hjdivad and
myself, allowing for Enumerology to use Ember's {array,reduce}Computed
properties, making enumerologies much more efficient.  Now, when there
are changes to the source collection, only the steps in the pipeline
with changes in their dependencies are fired, and then we try and use
add 1/subtract 1 semantics wherever possible.